requirements.txt
requests
python-whois
tldextract